+++ src/main/java/com/example/hotdeal/domain/event/domain/dto/EventResponse.java
@@ -1,5 +1,6 @@
-package com.example.hotdeal.domain.event.domain;
+package com.example.hotdeal.domain.event.domain.dto;
 
+import com.example.hotdeal.domain.event.domain.entity.Event;
 import com.example.hotdeal.domain.event.enums.EventType;
 import com.example.hotdeal.domain.product.product.domain.AddEventResponse;
 import lombok.Getter;
